refactor: consolidate workspace into single crate
Merge sikte-common and sikte-ebpf into main sikte crate while
preserving workspace structure for future expansion.

Changes:
- Move sikte-common types to sikte/src/common/
- Move C eBPF code from sikte-ebpf/src/ to sikte/ebpf-src/
- Update all imports from sikte_common:: to crate::common::
- Remove sikte-common and sikte-ebpf crate definitions
- Update build.rs to reference new ebpf-src/ paths
- Fix C header includes (add vmlinux.h to raw_trace_points.h)
- Add bytemuck derive feature (preserved from sikte-common)
- Update CLAUDE.md to reflect new single-crate architecture
- Move release profile to workspace level

The workspace remains with single member for easy future additions
(e.g., GUI crate).
