+++ Cargo.toml
@@ -1,7 +1,7 @@
 [workspace]
 resolver = "2"
-members = ["sikte", "sikte-common", "sikte-ebpf"]
-default-members = ["sikte", "sikte-common"]
+members = ["sikte"]
+default-members = ["sikte"]
 
 [workspace.package]
 license = "GNU Affero General Public License"
@@ -33,6 +33,6 @@ thiserror = { version = "2.0.16", default-features = false }
 tokio = { version = "1.40.0", default-features = false }
 which = { version = "6.0.0", default-features = false }
 
-[profile.release.package.sikte-ebpf]
+[profile.release]
 debug = 2
 codegen-units = 1
