--- sikte-common/Cargo.toml
@@ -1,16 +1,0 @@
-[package]
-name = "sikte-common"
-version = "0.1.0"
-edition.workspace = true
-
-license.workspace = true
-
-[features]
-default = []
-user = []
-
-[dependencies]
-bytemuck = { workspace = true, features = ["derive"] }
-
-[lib]
-path = "src/lib.rs"
