--- sikte-ebpf/Cargo.toml
@@ -1,13 +1,0 @@
-[package]
-name = "sikte-ebpf"
-version = "0.1.0"
-edition.workspace = true
-
-# C eBPF programs don't need Rust dependencies
-# Compilation handled by libbpf-cargo in sikte/build.rs
-
-[dependencies]
-# No dependencies needed for C eBPF code
-
-[build-dependencies]
-# No build dependencies needed
