--- sikte-ebpf/src/lib.rs
@@ -1,3 +1,0 @@
-#![no_std]
-
-// This file exists to enable the library target.
