+++ sikte/src/common/mod.rs
@@ -1,5 +1,3 @@
-#![no_std]
-
 pub mod constants;
 pub mod generic_types;
 pub mod raw_tracepoints;
