+++ sikte/src/ebpf/map_types.rs
@@ -1,6 +1,6 @@
 use libbpf_rs::MapCore;
 use libc::pid_t;
-use sikte_common::generic_types::Unused;
+use crate::common::generic_types::Unused;
 
 /// Syscall ring buffer wrapper
 pub struct SyscallRingBuf<'a> {
